Improve booking detail screen

+++ BookingSite/ViewModels/BookingWithTicketsViewModel.cs
@@ -6,7 +6,7 @@
         public DateTime CreatedAt { get; set; } = DateTime.Now;
         public string UserId { get; set; }
         public List<TicketPreviewViewModel> Tickets { get; set; } = new();
-        public decimal TotalPrice => Tickets.Sum(t => t.Price ?? 0);
+        public double TotalPrice => Tickets.Sum(t => t.Price ?? 0);
     }
 
     public class TicketPreviewViewModel
@@ -17,7 +17,7 @@
         public DateTime Date { get; set; }
         public string SeatClass { get; set; }
         public string MealDescription { get; set; }
-        public decimal? Price { get; set; }
+        public double? Price { get; set; }
         public bool IsCancelled { get; set; }
     }
 }
